Update G. Summation from 1 to N.cpp

+++ C++ Assignments/Codeforces/Sheet1/G. Summation from 1 to N.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-   int N,i,sum=0;
+   long long int N,i,sum=0;
    cin>>N;
    for(i=1; i<=N; i++)
    {
